requirements.txt
aiohttp==3.9.5
Django==5.0.4
django-boto==0.3.12
slack-sdk==3.27.1
python-dotenv==1.0.1
djangorestframework==3.15.1
boto3==1.34.84
requests==2.31.0
mysqlclient==2.2.4
django-mysql==4.12.0